optimize:Enhance HttpClient to support h2c
